Updated to work with KAI10100 CCD
Some significant differences in download formats
Different inquiry response
Different binning #'s
Different zone numbers for bin==1
There is still an outstanding issue where the server crashes (in cfitsio) immediately after ExposureComplete is called in grabImage.  However, it is not consistent.  Sometimes it will crash on every binning 1 frame (while binning 2 and 4 appear to never crash) and it will keep doing it.  Then it stops and will NOT replicate.  Puzzling.  Not calling dn->freeBuf appeared to stop the crash, but that may have been a coincedence.