Bayer change & diagnostics
Found the crash issue - at some point the frame got set to 2842 instead of 2840 and various bits of code were remembering that and resetting to it, which would then crash in cfitsio as it walked off the end of the buffer.
found an KAF8300 bug I introduced by reviews previous change, fixed.
With binning == 1, bayer is RGGB, but BGGR otherwise.